Modularize & deprecate formula helpers
